+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:slu_scav_hunt/firebase_options.dart';
 import 'package:slu_scav_hunt/theme.dart';
 import 'package:slu_scav_hunt/widgets/auth_gate.dart';
@@ -8,15 +9,23 @@ import 'package:slu_scav_hunt/providers/theme_provider.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
+  
+  // Initialize SharedPreferences
+  final sharedPreferences = await SharedPreferences.getInstance();
+
   // Initializes the Firebase App, typically called before any FlutterFire plugin is used
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
+  
   runApp(
     // Wraps app in a widget that stores the state of the provider 
     // Called before riverpod plugins
-    const ProviderScope(
-      child: MyApp(),
+    ProviderScope(
+      overrides: [
+        sharedPreferencesProvider.overrideWithValue(sharedPreferences),
+      ],
+      child: const MyApp(),
     ),
   );
 }